+++ test/Finbuckle.MultiTenant.AspNetCore.Test/Extensions/MultiTenantBuilderExtensionsShould.cs
@@ -11,6 +11,7 @@ using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Authentication.OpenIdConnect;
 using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using Moq;
@@ -21,14 +22,16 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
 {
     public class MultiTenantBuilderExtensionsShould
     {
-        public class TestTenantInfo : ITenantInfo
+        private class TestTenantInfo : ITenantInfo
         {
             public Guid Id { get; set; }
             public string? Identifier { get; set; }
             public string? Name { get; set; }
             public string? ConnectionString { get; set; }
+
             public string? ChallengeScheme { get; set; }
-            public string? CookiePath { get; set; }
+
+            // public string? CookiePath { get; set; }
             public string? CookieLoginPath { get; set; }
             public string? CookieLogoutPath { get; set; }
             public string? CookieAccessDeniedPath { get; set; }
@@ -48,8 +51,10 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             var sp = services.BuildServiceProvider();
 
             // Fake a resolved tenant
-            var mtc = new MultiTenantContext<TenantInfo>();
-            mtc.TenantInfo = new TenantInfo { Identifier = "abc" };
+            var mtc = new MultiTenantContext<TenantInfo>
+            {
+                TenantInfo = new TenantInfo { Identifier = "abc" }
+            };
             sp.GetRequiredService<IMultiTenantContextAccessor<TenantInfo>>().MultiTenantContext = mtc;
 
             // Trigger the ValidatePrincipal event
@@ -58,7 +63,8 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             httpContextMock.Setup(c => c.Items).Returns(new Dictionary<object, object?>());
             var scheme = sp.GetRequiredService<IAuthenticationSchemeProvider>()
                 .GetSchemeAsync(CookieAuthenticationDefaults.AuthenticationScheme).Result;
-            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>().Get(CookieAuthenticationDefaults.AuthenticationScheme);
+            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>()
+                .Get(CookieAuthenticationDefaults.AuthenticationScheme);
             var principal = new ClaimsPrincipal(new ClaimsIdentity());
             var authTicket = new AuthenticationTicket(principal, CookieAuthenticationDefaults.AuthenticationScheme);
             authTicket.Properties.Items[Constants.TenantToken] = "abc";
@@ -91,8 +97,10 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
 
 
             // Fake a resolved tenant
-            var mtc = new MultiTenantContext<TenantInfo>();
-            mtc.TenantInfo = new TenantInfo { Identifier = "abc" };
+            var mtc = new MultiTenantContext<TenantInfo>
+            {
+                TenantInfo = new TenantInfo { Identifier = "abc" }
+            };
             sp.GetRequiredService<IMultiTenantContextAccessor<TenantInfo>>().MultiTenantContext = mtc;
 
             // Trigger the ValidatePrincipal event
@@ -101,7 +109,8 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             httpContextMock.Setup(c => c.Items).Returns(new Dictionary<object, object?>());
             var scheme = sp.GetRequiredService<IAuthenticationSchemeProvider>()
                 .GetSchemeAsync(CookieAuthenticationDefaults.AuthenticationScheme).Result;
-            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>().Get(CookieAuthenticationDefaults.AuthenticationScheme);
+            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>()
+                .Get(CookieAuthenticationDefaults.AuthenticationScheme);
             var principal = new ClaimsPrincipal(new ClaimsIdentity());
             var authTicket = new AuthenticationTicket(principal, CookieAuthenticationDefaults.AuthenticationScheme);
             authTicket.Properties.Items[Constants.TenantToken] = "abc";
@@ -114,7 +123,7 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
         }
 
         [Fact]
-        public void PassprincipalValidationIfTenantMatch()
+        public void PassPrincipalValidationIfTenantMatch()
         {
             var services = new ServiceCollection();
             services.AddLogging();
@@ -125,8 +134,10 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
 
 
             // Fake a resolved tenant
-            var mtc = new MultiTenantContext<TenantInfo>();
-            mtc.TenantInfo = new TenantInfo { Identifier = "abc" };
+            var mtc = new MultiTenantContext<TenantInfo>
+            {
+                TenantInfo = new TenantInfo { Identifier = "abc" }
+            };
             sp.GetRequiredService<IMultiTenantContextAccessor<TenantInfo>>().MultiTenantContext = mtc;
 
             // Trigger the ValidatePrincipal event
@@ -135,7 +146,8 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             httpContextMock.Setup(c => c.Items).Returns(new Dictionary<object, object?>());
             var scheme = sp.GetRequiredService<IAuthenticationSchemeProvider>()
                 .GetSchemeAsync(CookieAuthenticationDefaults.AuthenticationScheme).Result;
-            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>().Get(CookieAuthenticationDefaults.AuthenticationScheme);
+            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>()
+                .Get(CookieAuthenticationDefaults.AuthenticationScheme);
             var principal = new ClaimsPrincipal(new ClaimsIdentity());
             var authTicket = new AuthenticationTicket(principal, CookieAuthenticationDefaults.AuthenticationScheme);
             authTicket.Properties.Items[Constants.TenantToken] = "abc";
@@ -161,19 +173,24 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             var sp = services.BuildServiceProvider();
 
             // Fake a resolved tenant
-            var mtc = new MultiTenantContext<TenantInfo>();
-            mtc.TenantInfo = new TenantInfo { Identifier = "abc1" };
+            var mtc = new MultiTenantContext<TenantInfo>
+            {
+                TenantInfo = new TenantInfo { Identifier = "abc1" }
+            };
             sp.GetRequiredService<IMultiTenantContextAccessor<TenantInfo>>().MultiTenantContext = mtc;
 
             // Trigger the ValidatePrincipal event
             var httpContextMock = new Mock<HttpContext>();
             httpContextMock.Setup(c => c.RequestServices).Returns(sp);
-            var httpContextItems = new Dictionary<object, object?>();
-            httpContextItems[$"{Constants.TenantToken}__bypass_validate_principal__"] = true;
+            var httpContextItems = new Dictionary<object, object?>
+            {
+                [$"{Constants.TenantToken}__bypass_validate_principal__"] = true
+            };
             httpContextMock.Setup(c => c.Items).Returns(httpContextItems);
             var scheme = sp.GetRequiredService<IAuthenticationSchemeProvider>()
                 .GetSchemeAsync(CookieAuthenticationDefaults.AuthenticationScheme).Result;
-            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>().Get(CookieAuthenticationDefaults.AuthenticationScheme);
+            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>()
+                .Get(CookieAuthenticationDefaults.AuthenticationScheme);
             var principal = new ClaimsPrincipal(new ClaimsIdentity());
             var authTicket = new AuthenticationTicket(principal, CookieAuthenticationDefaults.AuthenticationScheme);
             authTicket.Properties.Items[Constants.TenantToken] = "abc2";
@@ -209,12 +226,15 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             // Trigger the ValidatePrincipal event
             var httpContextMock = new Mock<HttpContext>();
             httpContextMock.Setup(c => c.RequestServices).Returns(sp);
-            var httpContextItems = new Dictionary<object, object?>();
-            httpContextItems[$"{Constants.TenantToken}__bypass_validate_principal__"] = true;
+            var httpContextItems = new Dictionary<object, object?>
+            {
+                [$"{Constants.TenantToken}__bypass_validate_principal__"] = true
+            };
             httpContextMock.Setup(c => c.Items).Returns(httpContextItems);
             var scheme = sp.GetRequiredService<IAuthenticationSchemeProvider>()
                 .GetSchemeAsync(CookieAuthenticationDefaults.AuthenticationScheme).Result;
-            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>().Get(CookieAuthenticationDefaults.AuthenticationScheme);
+            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>()
+                .Get(CookieAuthenticationDefaults.AuthenticationScheme);
             var principal = new ClaimsPrincipal(new ClaimsIdentity());
             var authTicket = new AuthenticationTicket(principal, CookieAuthenticationDefaults.AuthenticationScheme);
             authTicket.Properties.Items[Constants.TenantToken] = "abc2";
@@ -239,8 +259,10 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
 
 
             // Fake a resolved tenant
-            var mtc = new MultiTenantContext<TenantInfo>();
-            mtc.TenantInfo = new TenantInfo { Identifier = "abc1" };
+            var mtc = new MultiTenantContext<TenantInfo>
+            {
+                TenantInfo = new TenantInfo { Identifier = "abc1" }
+            };
             sp.GetRequiredService<IMultiTenantContextAccessor<TenantInfo>>().MultiTenantContext = mtc;
 
             // Trigger the ValidatePrincipal event
@@ -249,7 +271,8 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             httpContextMock.Setup(c => c.Items).Returns(new Dictionary<object, object?>());
             var scheme = sp.GetRequiredService<IAuthenticationSchemeProvider>()
                 .GetSchemeAsync(CookieAuthenticationDefaults.AuthenticationScheme).Result;
-            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>().Get(CookieAuthenticationDefaults.AuthenticationScheme);
+            var options = sp.GetRequiredService<IOptionsMonitor<CookieAuthenticationOptions>>()
+                .Get(CookieAuthenticationDefaults.AuthenticationScheme);
             var principal = new ClaimsPrincipal(new ClaimsIdentity());
             var authTicket = new AuthenticationTicket(principal, CookieAuthenticationDefaults.AuthenticationScheme);
             authTicket.Properties.Items[Constants.TenantToken] = "abc2";
@@ -278,7 +301,9 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             var schemeProvider = sp.GetRequiredService<IAuthenticationSchemeProvider>(); // Throws if fails
             Assert.IsType<MultiTenantAuthenticationSchemeProvider>(schemeProvider);
 
-            var strategy = sp.GetServices<IMultiTenantStrategy>().Where(s => s.GetType() == typeof(RemoteAuthenticationCallbackStrategy)).Single();
+            var strategy = sp
+                .GetServices<IMultiTenantStrategy>()
+                .Single(s => s.GetType() == typeof(RemoteAuthenticationCallbackStrategy));
             Assert.NotNull(strategy);
         }
 
@@ -310,7 +335,9 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
                 .WithRemoteAuthenticationCallbackStrategy();
             var sp = services.BuildServiceProvider();
 
-            var strategy = sp.GetServices<IMultiTenantStrategy>().Where(s => s.GetType() == typeof(RemoteAuthenticationCallbackStrategy)).Single();
+            var strategy = sp
+                .GetServices<IMultiTenantStrategy>()
+                .Single(s => s.GetType() == typeof(RemoteAuthenticationCallbackStrategy));
             Assert.NotNull(strategy);
         }
 
@@ -363,9 +390,36 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
         }
 
         [Fact]
+        public void ConfigurePerTenantAuthenticationConventions_UseDefaultChallengeSchemeOptionsIfNoTenantProp()
+        {
+            var services = new ServiceCollection();
+            services.AddOptions();
+            var defaultValue = "defaultScheme";
+            services.AddAuthentication(o => o.DefaultChallengeScheme = defaultValue)
+                .AddCookie()
+                .AddOpenIdConnect("defaultScheme", null!);
+            services.AddMultiTenant<TenantInfo>()
+                .WithPerTenantAuthenticationConventions();
+            var sp = services.BuildServiceProvider();
+
+            var ti1 = new TenantInfo
+            {
+                Id = "id1",
+                Identifier = "identifier1"
+            };
+
+            var accessor = sp.GetRequiredService<IMultiTenantContextAccessor<TenantInfo>>();
+            accessor.MultiTenantContext = new MultiTenantContext<TenantInfo> { TenantInfo = ti1 };
+
+            var options = sp.GetRequiredService<IAuthenticationSchemeProvider>();
+            Assert.Equal(defaultValue, options.GetDefaultChallengeSchemeAsync()!.Result!.Name);
+        }
+
+        [Fact]
         public void ConfigurePerTenantAuthentication_UseOpenIdConnectConvention()
         {
             var services = new ServiceCollection();
+            services.AddSingleton<IConfiguration>((new ConfigurationBuilder()).Build()); // net7.0+
             services.AddOptions();
             services.AddAuthentication().AddOpenIdConnect();
             services.AddMultiTenant<TestTenantInfo>()
@@ -384,7 +438,8 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             var accessor = sp.GetRequiredService<IMultiTenantContextAccessor<TestTenantInfo>>();
             accessor.MultiTenantContext = new MultiTenantContext<TestTenantInfo> { TenantInfo = ti1 };
 
-            var options = sp.GetRequiredService<IOptionsSnapshot<OpenIdConnectOptions>>().Get(OpenIdConnectDefaults.AuthenticationScheme);
+            var options = sp.GetRequiredService<IOptionsSnapshot<OpenIdConnectOptions>>()
+                .Get(OpenIdConnectDefaults.AuthenticationScheme);
 
             Assert.Equal(ti1.OpenIdConnectAuthority, options.Authority);
             Assert.Equal(ti1.OpenIdConnectClientId, options.ClientId);
@@ -395,6 +450,7 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
         public void ConfigurePerTenantAuthenticationConventions_UseOpenIdConnectConvention()
         {
             var services = new ServiceCollection();
+            services.AddSingleton<IConfiguration>((new ConfigurationBuilder()).Build()); // net7.0+
             services.AddOptions();
             services.AddAuthentication().AddOpenIdConnect();
             services.AddMultiTenant<TestTenantInfo>()
@@ -413,11 +469,45 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             var accessor = sp.GetRequiredService<IMultiTenantContextAccessor<TestTenantInfo>>();
             accessor.MultiTenantContext = new MultiTenantContext<TestTenantInfo> { TenantInfo = ti1 };
 
-            var options = sp.GetRequiredService<IOptionsSnapshot<OpenIdConnectOptions>>().Get(OpenIdConnectDefaults.AuthenticationScheme);
+            var options = sp.GetRequiredService<IOptionsSnapshot<OpenIdConnectOptions>>()
+                .Get(OpenIdConnectDefaults.AuthenticationScheme);
 
             Assert.Equal(ti1.OpenIdConnectAuthority, options.Authority);
             Assert.Equal(ti1.OpenIdConnectClientId, options.ClientId);
             Assert.Equal(ti1.OpenIdConnectClientSecret, options.ClientSecret);
+        }
+
+        [Fact]
+        public void ConfigurePerTenantAuthenticationConventions_UseDefaultOpenIdConnectOptionsIfNoTenantProp()
+        {
+            var services = new ServiceCollection();
+            services.AddSingleton<IConfiguration>((new ConfigurationBuilder()).Build()); // net7.0+
+            var defaultValue = "https://defaultValue";
+            services.AddOptions().AddAuthentication()
+                .AddOpenIdConnect(options =>
+                {
+                    options.Authority = defaultValue;
+                    options.ClientId = defaultValue;
+                    options.ClientSecret = defaultValue;
+                });
+            services.AddMultiTenant<TenantInfo>()
+                .WithPerTenantAuthenticationConventions();
+            var sp = services.BuildServiceProvider();
+
+            var ti1 = new TenantInfo
+            {
+                Id = "id1",
+                Identifier = "identifier1"
+            };
+
+            var accessor = sp.GetRequiredService<IMultiTenantContextAccessor<TenantInfo>>();
+            accessor.MultiTenantContext = new MultiTenantContext<TenantInfo> { TenantInfo = ti1 };
+
+            var options = sp.GetRequiredService<IOptionsSnapshot<OpenIdConnectOptions>>()
+                .Get(OpenIdConnectDefaults.AuthenticationScheme);
+            Assert.Equal(defaultValue, options.Authority);
+            Assert.Equal(defaultValue, options.ClientId);
+            Assert.Equal(defaultValue, options.ClientSecret);
         }
 
         [Fact]
@@ -442,7 +532,8 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             var accessor = sp.GetRequiredService<IMultiTenantContextAccessor<TestTenantInfo>>();
             accessor.MultiTenantContext = new MultiTenantContext<TestTenantInfo> { TenantInfo = ti1 };
 
-            var options = sp.GetRequiredService<IOptionsSnapshot<CookieAuthenticationOptions>>().Get(CookieAuthenticationDefaults.AuthenticationScheme);
+            var options = sp.GetRequiredService<IOptionsSnapshot<CookieAuthenticationOptions>>()
+                .Get(CookieAuthenticationDefaults.AuthenticationScheme);
 
             Assert.Equal(ti1.CookieLoginPath, options.LoginPath);
             Assert.Equal(ti1.CookieLogoutPath, options.LogoutPath);
@@ -471,11 +562,44 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             var accessor = sp.GetRequiredService<IMultiTenantContextAccessor<TestTenantInfo>>();
             accessor.MultiTenantContext = new MultiTenantContext<TestTenantInfo> { TenantInfo = ti1 };
 
-            var options = sp.GetRequiredService<IOptionsSnapshot<CookieAuthenticationOptions>>().Get(CookieAuthenticationDefaults.AuthenticationScheme);
+            var options = sp.GetRequiredService<IOptionsSnapshot<CookieAuthenticationOptions>>()
+                .Get(CookieAuthenticationDefaults.AuthenticationScheme);
 
             Assert.Equal(ti1.CookieLoginPath, options.LoginPath);
             Assert.Equal(ti1.CookieLogoutPath, options.LogoutPath);
             Assert.Equal(ti1.CookieAccessDeniedPath, options.AccessDeniedPath);
+        }
+
+        [Fact]
+        public void ConfigurePerTenantAuthenticationConventions_UseDefaultCookieOptionsIfNoTenantProp()
+        {
+            var services = new ServiceCollection();
+            var defaultValue = "/defaultValue";
+            services.AddOptions().AddAuthentication().AddCookie(options =>
+            {
+                options.LoginPath = defaultValue;
+                options.LogoutPath = defaultValue;
+                options.AccessDeniedPath = defaultValue;
+            });
+            services.AddMultiTenant<TenantInfo>()
+                .WithPerTenantAuthenticationConventions();
+            var sp = services.BuildServiceProvider();
+
+            var ti1 = new TenantInfo
+            {
+                Id = "id1",
+                Identifier = "identifier1"
+            };
+
+            var accessor = sp.GetRequiredService<IMultiTenantContextAccessor<TenantInfo>>();
+            accessor.MultiTenantContext = new MultiTenantContext<TenantInfo> { TenantInfo = ti1 };
+
+            var options = sp.GetRequiredService<IOptionsSnapshot<CookieAuthenticationOptions>>()
+                .Get(CookieAuthenticationDefaults.AuthenticationScheme);
+
+            Assert.Equal(defaultValue, options.LoginPath);
+            Assert.Equal(defaultValue, options.LogoutPath);
+            Assert.Equal(defaultValue, options.AccessDeniedPath);
         }
 
         [Fact]
@@ -498,7 +622,7 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             var strategy = sp.GetRequiredService<IMultiTenantStrategy>();
             Assert.IsType<BasePathStrategy>(strategy);
         }
-        
+
         [Fact]
         public void AddBasePathStrategyDefaultRebaseFalse()
         {
@@ -513,7 +637,7 @@ namespace Finbuckle.MultiTenant.AspNetCore.Test.Extensions
             var options = sp.GetRequiredService<IOptions<BasePathStrategyOptions>>();
             Assert.False(options.Value.RebaseAspNetCorePathBase);
         }
-        
+
         [Fact]
         public void AddBasePathStrategyWithOptions()
         {